// types/Tochki.ts
export interface IAppTochki {
    id: number;
    elm___id: string;
    elm___name: string;
    elm___createdAt: string;
    elm___updatedAt: string;
    elm_akciya: string;
    elm_apt_id: string;
    elm_municipalnyi_raion: string;
    elm_adres_apteki: string;
    elm_kategoriya: string;
    elm_set: string;
    elm_yuridicheskoe_lico: string;
    elm_inn: string;
    elm_telefon: string;
    elm_forma_vykladki: string;
    elm_geokoordinaty: string;
    elm_rekomendovannoe_vremya_posesheniya: string;
    elm_uronekst_sashe_7: string;
    elm_po_tab_12_mg_10: string;
    elm_po_supp_12_mg_10: string;
    elm_po_supp_6_mg_10: string;
    elm_po_rr_6_mg_5: string;
    elm_lg_supp_10: string;
    elm_lg_supp_20: string;
    elm_lg_fl_5: string;
    elm_artneo_kaps_30: string;
    elm_velson_tab_30: string;
    elm_vitaferr_kaps_30: string;
    elm_df_kapli_15_ml: string;
    elm_omegika_kaps_30: string;
    elm_im_krem_30_g: string;
    elm_vitaferr_ekspress: string;
    elm_uronekst_sashe_14: string;
    elm_artneo_akciya_na_pokupatelya: string;
    elm_artneo_prochie_aktivnosti: string;
    elm_velson_akciya_na_pokupatelya: string;
    elm_velson_prochie_aktivnosti: string;
    elm_vitaferr_akciya_na_pokupatelya: string;
    elm_vitaferr_prochie_aktivnosti: string;
    elm_demefecil_akciya_na_pokupatelya: string;
    elm_demefecil_prochie_aktivnosti: string;
    elm_imoferaza_akciya_na_pokupatelya: string;
    elm_imoferaza_prochie_aktivnosti: string;
    elm_po_dmf_akciya_na_pokupatelya: string;
    elm_po_dmf_prochie_aktivnosti: string;
    elm_po_vrachebnye_formy_akciya_na_pokupatelya: string;
    elm_po_vrachebnye_formy_prochie_aktivnosti: string;
    elm_po_vse_skyu_akciya_na_pokupatelya: string;
    elm_po_vse_skyu_prochie_aktivnosti: string;
    elm_po_tab_akciya_na_pokupatelya: string;
    elm_po_tab_prochie_aktivnosti: string;
    elm_polioksidonii_akciya_na_pokupatelya: string;
    elm_polioksidonii_prochie_aktivnosti: string;
    elm_uronekst_akciya_na_pokupatelya: string;
    elm_uronekst_prochie_aktivnosti: string;
    elm_po_rr_akciya_na_pokupatelya: string;
    elm_po_rr_prochie_aktivnosti: string;
    elm_po_supp_12_akciya_na_pokupatelya: string;
    elm_po_supp_12_prochie_aktivnosti: string;
    elm_po_supp_6_akciya_na_pokupatelya: string;
    elm_po_supp_6_prochie_aktivnosti: string;
    elm_sandevit_akciya_na_pokupatelya: string;
    elm_sandevit_prochie_aktivnosti: string;
    elm_velson_30_akciya_na_pokupatelya: string;
    elm_velson_30_prochie_aktivnosti: string;
    elm_omegika_akciya_na_pokupatelya: string;
    elm_omegika_prochie_aktivnosti: string;
    elm_uronekst_7_akciya_na_pokupatelya: string;
    elm_uronekst_7_prochie_aktivnosti: string;
    elm_uronekst_7_i_14_akciya_na_pokupatelya: string;
    elm_uronekst_7_i_14_prochie_aktivnosti: string;
    elm_podioksidonii_tab_demefecil_kap_akciya_na_pokupatelya: string;
    elm_podioksidonii_tab_demefecil_kap_prochie_aktivnosti: string;
    vizit__id: string;
    vizit_status: string;
    vizit_kommentarii: string;
    vizit_data_vremya: string;
    vizit_data_zakrytiya: string;
}