// types/Vizity.ts
export interface IAppVizityAdd {
    elm_tochka: string;
    elm_data_vremya: string;
    elm_postponed: string;
}